fixed issue with auth
Also added a memory cache to the auth process, to store user_ids to save us from having to look the users up in the database for every request.

If the user is not in the cache, the database will be queried and the user id will be added to the cache.

If the user is not in the database, it is the first time the user signs in and the user will be created in the database and stored in the cache.